+++ examples/vite-plugin-ssr/pages/index/index.page.tsx
@@ -6,7 +6,7 @@ export { Page };
 function Page() {
 	return (
 		<>
-			<h1>Welcome</h1>
+			<h1>Hello from vite-plugin-ssr</h1>
 			This page is:
 			<ul>
 				<li>Rendered to HTML.</li>
